+++ nextentity-examples/src/main/java/io/github/nextentity/examples/repository/IExtendedEntityRepository.java
@@ -1,0 +1,6 @@
+package io.github.nextentity.examples.repository;
+
+import io.github.nextentity.spring.Repository;
+
+public interface IExtendedEntityRepository<T, ID> extends Repository<T, ID> {
+}